+++ packages/werkbank/lib/src/addons/src/accessibility/src/_internal/semantics_inspector/node_info/semantics_inspector_node_info.dart
@@ -150,6 +150,16 @@ class SemanticsInspectorNodeInfo extends StatelessWidget {
         ', ',
       ),
     );
+    addRawStringField(
+      'custom actions',
+      [
+        for (final customActionId
+            in data.customSemanticsActionIds ?? const Iterable<int>.empty())
+          CustomSemanticsAction.getAction(customActionId)?.label,
+      ].nonNulls.join(
+        ', ',
+      ),
+    );
 
     return fields;
   }