+++ packages/werkbank/CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add advanced search features. TL;DR: `fuzzy text`, `<field>:fuzzy text`, `"precise text"`, `<field>:"precise text"`
+
 # 0.15.0
 
 - Initial release