+++ packages/werkbank/CHANGELOG.md
@@ -5,6 +5,7 @@
     - How the descriptions are merged can be controlled by the new `mergeStrategy` parameter.
   - Added `forUseCase` parameter for `c.description(...)` to specify whether the description is for the use case or the current node.
   - The name of a `WerkbankFolder` or `WerkbankComponent` is now displayed above its description.
+- Fix that semantics inspector in "Inspection" mode would not allow to select semantic boxes in the main view on web.
 
 # 0.15.0